refactor : 숫자 생성 클래스 패키지 분리 및 로또 숫자 관련 상수 분리
