+++ src/main/java/constant/LottoConstant.java
@@ -1,0 +1,8 @@
+package constant;
+
+public class LottoConstant {
+	public static final int MIN_NUMBER = 1;
+	public static final int MAX_NUMBER = 45;
+
+	public static final int NUMBER_OF_NUMBERS = 6;
+}
