--- src/main/java/domain/LottoNumbersGenerator.java
@@ -1,8 +1,0 @@
-package domain;
-
-import java.util.List;
-
-public interface LottoNumbersGenerator {
-
-	List<Integer> generate();
-}
