+++ src/main/java/domain/Lottos.java
@@ -1,5 +1,7 @@
 package domain;
 
+import util.LottoNumbersGenerator;
+
 import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
