+++ src/main/java/util/LottoNumbersGenerator.java
@@ -1,0 +1,8 @@
+package util;
+
+import java.util.List;
+
+public interface LottoNumbersGenerator {
+
+	List<Integer> generate();
+}
