+++ src/test/java/LottosTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import util.RandomLottoNumberGenerator;
 
 public class LottosTest {
 	static Lottos lottos;
